Move util functions to better places
