+++ src/modules/deps/list.ts
@@ -1,19 +1,8 @@
 import log from '../../logger'
 import {apps} from '../../clients'
+import {removeNpm} from './utils'
 
 const {getDependencies} = apps
-const isNpm = dep => dep.startsWith('npm:')
-export const removeNpm = (deps, inValues?) => {
-  Object.keys(deps).forEach(key => {
-    if (isNpm(key)) {
-      return delete deps[key]
-    }
-    if (inValues) {
-      deps[key] = deps[key].filter(d => !isNpm(d))
-    }
-  })
-  return deps
-}
 
 export default {
   description: 'List your workspace dependencies',
