+++ src/modules/deps/utils.ts
@@ -1,0 +1,13 @@
+const isNpm = dep => dep.startsWith('npm:')
+
+export const removeNpm = (deps, inValues?) => {
+  Object.keys(deps).forEach(key => {
+    if (isNpm(key)) {
+      return delete deps[key]
+    }
+    if (inValues) {
+      deps[key] = deps[key].filter(d => !isNpm(d))
+    }
+  })
+  return deps
+}
